override signals table default padding
